Add pc2 template for scimbal cam and rename pc1 scimbal cam
